Move Javafx renderer implementations into own package
