+++ jsvg-javafx/src/main/java/com/github/weisj/jsvg/ui/jfx/renderer/FXSVGRendererJavaFX.java
@@ -19,14 +19,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *
  */
-package com.github.weisj.jsvg.ui.jfx.skin;
+package com.github.weisj.jsvg.ui.jfx.renderer;
 
-import java.awt.*;
-import java.awt.image.BufferedImage;
-import javafx.embed.swing.SwingFXUtils;
 import javafx.scene.Node;
-import javafx.scene.image.ImageView;
-import javafx.scene.image.WritableImage;
+import javafx.scene.canvas.Canvas;
+import javafx.scene.canvas.GraphicsContext;
+import javafx.scene.effect.BlendMode;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -34,84 +32,61 @@ import org.jetbrains.annotations.Nullable;
 import com.github.weisj.jsvg.SVGDocument;
 import com.github.weisj.jsvg.renderer.NullPlatformSupport;
 import com.github.weisj.jsvg.renderer.animation.AnimationState;
-import com.github.weisj.jsvg.renderer.jfx.impl.bridge.FXRenderingHintsUtil;
+import com.github.weisj.jsvg.renderer.jfx.impl.FXOutput;
 import com.github.weisj.jsvg.renderer.output.Output;
 import com.github.weisj.jsvg.ui.jfx.FXSVGCanvas;
 import com.github.weisj.jsvg.view.FloatSize;
 import com.github.weisj.jsvg.view.ViewBox;
 
-class FXSVGRendererAWT implements FXSVGRenderer {
+public final class FXSVGRendererJavaFX implements FXSVGRenderer {
+    private final Canvas canvas;
+    private final GraphicsContext graphics;
 
-    private ImageView fxImageView;
-
-    private BufferedImage awtImage;
-    private WritableImage fxImage;
-    private int currentRTWidth = -1;
-    private int currentRTHeight = -1;
-
-    public FXSVGRendererAWT() {
-        fxImageView = new ImageView();
-    }
-
-    private void setupRenderTargets(int width, int height) {
-        awtImage = new BufferedImage(width, height, BufferedImage.TYPE_INT_ARGB);
-        fxImage = new WritableImage(width, height);
-        currentRTWidth = width;
-        currentRTHeight = height;
-        fxImageView.setImage(fxImage);
-    }
-
-    private void disposeRenderTargets() {
-        awtImage = null;
-        fxImage = null;
-        currentRTWidth = -1;
-        currentRTHeight = -1;
-    }
-
-    private void flush() {
-        fxImage = SwingFXUtils.toFXImage(awtImage, fxImage);
+    public FXSVGRendererJavaFX() {
+        canvas = new Canvas();
+        graphics = canvas.getGraphicsContext2D();
     }
 
     @Override
     public FXSVGCanvas.@NotNull RenderBackend getBackend() {
-        return FXSVGCanvas.RenderBackend.AWT;
+        return FXSVGCanvas.RenderBackend.JavaFX;
     }
 
     @Override
     public void render(@NotNull SVGDocument svgDocument, @Nullable ViewBox viewBox,
             @Nullable AnimationState animationState) {
-        FloatSize size = svgDocument.size();
-        int width = (int) size.width;
-        int height = (int) size.height;
+        FloatSize svgSize = svgDocument.size();
+        double width = svgSize.getWidth();
+        double height = svgSize.getHeight();
 
-        if (currentRTWidth != width || currentRTHeight != height) {
-            disposeRenderTargets();
-            setupRenderTargets(width, height);
+        if (canvas.getWidth() != width || canvas.getHeight() != height) {
+            canvas.setWidth(svgSize.getWidth());
+            canvas.setHeight(svgSize.getHeight());
         }
-        Graphics2D g = awtImage.createGraphics();
-        Output output = Output.createForGraphics(g);
-        FXRenderingHintsUtil.setupDefaultJFXRenderingHints(output);
-        g.setBackground(new Color(0, 0, 0, 0));
+
+        graphics.save();
         try {
-            g.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-            g.setRenderingHint(RenderingHints.KEY_STROKE_CONTROL, RenderingHints.VALUE_STROKE_PURE);
-            g.clearRect(0, 0, width, height);
+            graphics.setTransform(1, 0, 0, 1, 0, 0);
+            graphics.setGlobalAlpha(1D);
+            graphics.setGlobalBlendMode(BlendMode.SRC_OVER);
+            graphics.clearRect(0, 0, width, height);
+
+            Output output = FXOutput.createForGraphicsContext(graphics);
             svgDocument.renderWithPlatform(NullPlatformSupport.INSTANCE, output, viewBox, animationState);
+            output.dispose();
         } finally {
-            g.dispose();
+            graphics.restore();
         }
-        flush();
     }
 
     @Override
     public void dispose() {
-        disposeRenderTargets();
-        fxImageView = null;
+        // do nothing
     }
 
     @Override
     public @NotNull Node getFXNode() {
-        return fxImageView;
+        return canvas;
     }
 
 }
